MAJOR parallization update Overhead from firstprivate (OpenMP) removed First of all, the testing time for the given codes with parallization are 2.9e0, 7.44e-3, 4.4e-1. After parallization, the all the computing time increased dramatically because of overhead from firstprivate. The revised codes are using the parallization scheme without firstprivate and the shared with array of MATRIX class. On this regards, the testing time become 3.49e0, 7.73e-3, and 4.4e-1.
Note that the testing time without parallization is much faster: 1.0,
2.13e-2, 2.39e-1. This can be regarded as the problem for Monte-Carlo
steps. From the previous studies, it is already checked the 80% of
computing time is related with the computing pdf from other chain
end. That is the most tricky parts to parallization scheme applicable on
here.

So, various jobs left on here.