Removing MC overhead for PDF map One time computing all PDF map for beads In previously, the distance of pairs are computed only for the needed index. That seems to be reasonable, but has lots of overheads. Consider M_i as MC steps for i-th evolution time, then the cost to get PDF maps for a given bead index is O(Np) where Np is number of particle. Note that the different computing time between beads are negligible, say TP. The overall computing for PDF map becomes TP*sum_i M_i.
Since the configurational information is not changed during MC steps, we
can compute PDF maps between all existing pairs. That is all the pairs
computed twice time, which is also the case for overhead. Even if this
worst case, the overall time becomes TP*Nt*Np, where Nt is number of
times for Langevin dynamics. Note that M_i typically around thousands
order for testing purpose while the Np on this case is around 40.

Here, one more information is needed to classify for 3-dimensional
computing - the relation between Np and M_i(Np) is not classified up to
now -. Therefore, the time analysis on here only for the references.