DEBUG: Memory Leak is fixed Memory leak from MATRIX and MATRIX_LONG classes Originally, the problem is come from the initial() function for both of classes. This is because the lack of checking the existing allocated data. Without caution, when we call the method ".initial", it will allocate new array to the data pointer. That means the existing memory is just left for the system without any re-use of it. This make limitation to computing since the all memory occupied by this memory leakage.
To prevent memory overfloor, the Mac OS X is seems to use two
scheme: (i) saving the memoryfile to virtual memory that is SSD for my
mac and (ii) killing the files with error message "Killed:
9". Therefore, the memory history increases almost lineary due to time,
then sudden drop by the first scheme and this virtual memory is
represented by inactive. Due to further memory leakage, the active
memory still increase then it meet the second preventing
scheme. Therefore, even if the program is properly worked and the data
is not affected by the leakage, the computing will be killed by OS in
certain time. On this regards, this debugging is of importance to
prevent the potential problem.