TRACKING PROBLEM: -ipo option Interprocedure optimization for files has problem The IPO option supported by Intel Compiler (icpc) said the problem with the code. At the moment, I have no sure where the problem comes from. Hence, -ipo option is disabled for temporally, but in near future, it will update with proper ways.
One doubt is related with the problem is that 'virtual' option for other
library. However, the simple removing the option is not really working,
and during the revision of source code, the infrastructure might be
corrupted. The main problem is related with memory removing which is not
allocated position. On this regards, the results of data is not affected
by the problem. Even if the source code is properly working without -ipo
option, it would be better to track for this case to be safe side.