PERFORMANCE_UPDATE:Major bisection, Minor MKL 11 faster than the previous codes The major performance update is related with the searching algorithm for new chain ends, which is originally set by backtrace from the higher probability to low probability.
Before describe the bisection, it would
be better to explain how original backtrace works: given CDF is already
sorted, monotonically increasing function with respect to index i, and
many of the probability reaching 0. Therefore, the backtrace of
monotonically increasing probability map is quite efficiency. There is
the other reason for this efficiency: we do not have to compute
CDF map for every MC steps because the CDF only depends on the
trajectory which is enough to compute only once a each time. During MC
steps, it will be re-used which typically computed over 10000
times. Typical finding of backtrace is around 150 steps from its
initials when number of particles are 640.

The bisection algorithm is much efficiency since our CDF is
monotonically increasing function. For given number of particles, 640,
the ideal bisection will counting int(log_2(640)) = 9. However, the
domain for this bisection is in integer rather than real number, there
is missing of the counting, 5/2 = 2 (in integer for c++). So, the
compensation is computed by conditional phrase (dk%2 != 0) that will
return 0 when modulo 2 becomes zero (even number) while 1 when odd
numbers are given. In this case, the bisection step is computed as
int(log_2(Np)) + #given_odd_counting ~ 11 when Np=640.

The performance dramatically increasing, since the MC step occupied over
99% of computation time and this backtrace occupied 66% of MC step. The
number of steps to find the new chain ends is 14 times faster than
existing one which contribute total computation time is 11 times faster
than the previous one.