IMPORTANT:COMPATIBLE INFRA STRUCTURE WITH MKL BLAS MKL BLAS supporting The motivation is related with the fact that the time consuming is focused on the area of updating after previous refinement (from backtrace to bisection). For implementation of cblas via MKL, the infrastructure of association (and connectivity) classes should be changed. One main point is that the number of beads are separated in the way of mkl_malloc rather than matrix shape. So, the structure of the object from these class has the form of HASH[i](j) rather than HASH(i,j). On this regards, HASH[i] will return column vector with rank of cond_longer.
This change is critical for the interface design of each functions;
therefore, all the related interface have been changed. One of the
important change is the HASH and weight matrix changed from MATRIX_LONG
to MATRIX which posses double types for each elements. This is due to
the fact that the BLAS interface taking the double* argument rather than
LONG* argument. I have no sure it is possible to use type conversion
by (double*) from (MKL_LONG*). In addition, many of hash table and
weight are used for multiplication with double type variables. From
these update, we can remove the dependency issues with the my own
package matrix_long_ed.h. Also printing out is related with the column
vector rather than row vector in previous approach, which make new
function for fprint_LONG_transpose.

Shortly, the infrastructure changes are related with the compatible form
with CBLAS supported by MKL. This changes are making array of column
vectors rather than matrix with has rows and columns. From this aspect,
the all the variables changed from row vector to column vector.