BRANCH:BUG_FIX:CHECK_DONE: Static Corruption Purely, the static corrpution is removed The original static corruption is related with the dynamically allocated MATRIX object inside OpenMP loop. This is already fixed in the main branch but it is buried with different change such as MKL_LONG to long change, which is no real necessary in this case. In addition, the latter changes on the master branch has the problematic based on the number of associations in comparison with the previous code, this debugging branch will be used for futher works.
The check is done using without '-openmp' option during compile, and
check the exactly the same results from the previous results. In
addition, the potential on this code is not general purpose but FENE
connector is used with given cut-off radius: RM_R0_ratio. Therefore, the
result is valuable in order to check the problem trace. Finally, no
difference was checked in the comparison with the previous code, which
means the master branch was problematic, while this DEBUG_9a1dc34_new
branch from 9a1dc34 point of the master branch is stable in terms of
results and codes.

There are left things existing such as generalization of the potential
maps and removing some overhead, which will be re-developed by
step-by-step and commit individually on this branch. If there is no
changes due to this re-building procedure, the master branch will be
merged by this fetch in order to provide stable version.