BRANCH:REMOVING UN-Necessary parts Removing some library, commented lines The fixed point is include parts that does not really used inside the code. The other aspect is related with the removing unnecessary commented line. Note that the MATRIX class is not re-builded since it include potentially useful code. The commented lines are buggy but good point for the direction of update.
The deletion of lin_alg library is due to the complicate of usage and
debugging of the code. The library itself is designed for interface
function with GSL library. However, handling GSL inside of MKL structure
already experienced MKL interface. On this regards, using new design
inside of the code is quite heavy to handle.

At this moment, the interface functions are related with MATRIX library
and RANDOM library that are based on MKL and GSL, respectively. The
reason is simple: it is used frequently inside the code, that need to
extract individual library with compatible interface design.