BRANCH:UPDATE:INDEX, TRUTH MAP IMPORTANT UPDATE for TRUTH_MAP Before going further, I will state some reference related with this revision: research note for 10 NOV 2015 and 24 NOV 2015, some of bug fix in the git log file 6f20271.
The original TRUTH_MAP approach have been used from its origin. I have
no sure why it is used rather than direct approach. It might be come
from some efficiency inside code at that time (not this time). However,
the basic approach on the previous code might be WRONG since it uses the
selection probability for chain from chains on index_itself while the
action is taken from the it's other chain ends. That means the DEL
action do delete the association which detach chain ends attached
index_itself rather than index_others. The ADD action will be the same,
but the MOV action also has different role. Hence, the source code is
revised based on the new action identifier where the main actor is the
chain attached index_itself.

The new TRUTH_MAP uses DEL as detach associated chain from index_other
and attach it to index_itself. The MOV action is taking detachment from the
index_other and attach it to index_new. So, the attachment at the
index_itself remains. The results of difference shows the basic
different number of associations. For instance, FENE with r11 case
in the previous code started with the 3706 number of associations while
the revised code started with 3555, which shows 200 associations is
deleted and the tendency shows quite shift manner at the initial parts,
but not exactly duplicate. So, we cannot assure it will be the same in
the equilibrium space in terms of number of associations. The further
structure changes during this simulation will be reported later parts.