BRANCH:UPDATE:BUGFIX:INDEX TREATEMENT Index handling is developed First of all, there is error message that pointer is being freed without allocated. This error message happens the only last section as the previous case (static corruption case). At the moment, the only clue is related with ASSOCIATION class which might violate the rule of three.
Since the index handling for the association/dissociation kinetics is of
importance to follow proper statistics, the index handling interface
have been updated. The results is exactly same with the previous
commited update, but different performance and handling code. So, this
revision is for improving the understandability without loss of performance.