BRANCH:UPDATE: compile option and refine comments Adding comments for the option of compile Note that -ipo (interprocedure optimization option) is turned on again that was problematic in the previous history. I guess that the static corruption might be the problem to handle -ipo option in previous days.
The IPO option is supported by MKL is related with the compiler to
analyze the code to determine the specific optimization which includeing
various aspect. For detail, see the Intel hompage about icpc compiler
options. Shortly, the IPO option optimization is targeting on the whole
line of structure rather than single functional line that benefits to
optimize the code in many aspects.