PARALLEL:INFRASTRUCTURE:without -openmp option The locking procedure properly working Note that the proposed scheme is applied without -openmp option, so this update is restricted with the case for the single processor. Even so, the updated code is worth to submit on this way in order to remove potential problems.
Before describe something, it would be mentioned that the comparison
without -openmp option have been tested with single manner of the
statistics that exactly matched with the previous code. This suggested
that there is NO violate in the case for statistics since the random
numbers are generated by deterministic way. Note that whenever we turned
on the -openmp option, we cannot compare between results step by step
because the parallelized MC steps (and also Brownian time evolution)
uses differently generated random-number generation. This is due to the
fact that gsl_rng is NOT thread-safety, which is needed to make
different stream in different thread. In this revised code, therefore,
several techniques are applied:
(i) local parallel for-loop inside while-loop for MC steps;
(ii) class LOCK is defined in lib_parallel;
(iii) IDX is dynamically allocated with N_THREADS;
(iv) r_boost is dynamically allocated with N_THREADS;

(i) First of all, the parallel scheme is applied using localize for-loop
inside while-loop. This is basically the concept for 'omp parallel for'
which benefit A LOT for OpenMP parallelization since OpenMP does not allow
to finish jobs inside parallel loop. To be specific, if we perform
parallel while-loop, the identifier for exit loop should be explicitly
defined before loop is started. It means we cannot adjust IDENTIFIER
inside loop based on EQUILIBRIUM JUDGMENT. To the my knowledge, there
are some tweaks exist on this aspect, which is seemed to be un-stable
and NOT recommended pathway for code. Hence, I have used nested loop:
while(IDENTIFIER) then parallelize for-loop inside. On this scheme, the
for-loop iterate at least the basic number of MC-steps using already
given condition, N_steps_block - the stabilizer for equilibrium
identifier -. At this moment, N_steps_block is defined on .inp file with
100 (as default) which suggesting that we are parallelize 100
for-loop. Therefore, the conditional phrase for while have changed.

(ii) class LOCK is the kernel of parallelization scheme that suggest the
way to check the bead is in-visited state or not. Therefore, not only
shared condition is applied, but the 'critical' directive is applied in
checking region. The 'critical' region is computed only for the single
thread while the other thread, if it is reached, will wait until
on-going computation will be finished. This benefit A LOT for identify
the beads is LOCKED or NOT since the critical region is relatively short
compared with other MC steps, and the the 'directive' region is only
applied on the critical region that makes tune the time sequence of
computation for all the threads (if the averaged computation time is the
same to each other). In reality, of course, the computation time for
each thread is difference due to the fact that difference actions:
CANCEL, ADD, DEL, MOV.

(iii) INDEX_MC is dynamically allocated with IDX_ARR[N_THREADS]. Inside
loop, this is re-used with reference variable: INDEX_MC &IDX =
IDX_ARR[given_thread_number]. So, inside the loop, many of index
treatment can be done through members of IDX.

(iv) gsl_rng *r_boost -> gsl_rng **r_boost_arr. Since gsl_rng is NOT
thread-safe, using one stream that shared between thread is NOT
recommended. Hence, the stream is generated with number of threads and
the stream is used with r_boost_arr[omp_get_thread_num()] inside the
loop. On this scheme, each stream will be used once in the parallel
regime that make to avoid the violate cases for random-number
generation. Potentially, change gsl_rng to Vector Statistics
Library (VSL, Intel) will be useful since VSL is thread-safe.