UPDATE: cell-list implementation (basic) The basic cell-list implementation is done The comments are not neutralization at this moment (to be on the safe side). There are several bug exists, and most of are related with the newaly introduced cell-list scheme. For instance, it frequently mistaken for N_cells and N_neighbor_cells where the former is N_div^N_dimension while the latter is N_sf^N_dimension.
For simplification, the equilibration option is disabled, so it would be
better to introduce equilibration option through new main function
defined.

There was a bug for cut-off scheme which will call
Modified_Boltzmann_Distribution but duplicate for scale factor. Hence,
the introduced cut-off scheme (because we are using cut-off as cell-list
without real cut-off for Boltzmann distribution which is not isotropic
since cell is made on squres). The bug is fixed at this moment, show the
same results compared with the test version.

The performance dramatically increases. The test condition is 400
micelles in 10^3 in dimensionless volume. The cut-off radius is set with
2.0 (checked without cut-off which shows that maximum bridges is around
1.8 and beyond 2.0 happens very occasionally), so the number of cells
per axis becomes 5, which implies the number of cells on the system is
5^3=125. In my Macbook, 4 processor is
used for parallel computation and it is 3.27 times faster than the
existing one that without cell-list. This factor is increasing function
with respect to box size which is eventual benefit for future works.


Former-commit-id: 94a6c9c4c8de697c7d175067654a3d5e1c7da5ed