[UPDATE] Post-Processing: Virials Computation virial have been re-visited There are problem for the definition of virial stress. For given pair-wise potential, the virial stress is given by the relative vector, Q_ij = R_j - R_i, rather than the position vector, $R_i$, individually, Typically, it is identical to sum over all connector vector and individual position vector divided by 2. With PBC, however, typical sum over RF tensor is not the same for QF when there are boundary change have happened. Hence, this revised post-processing uses connector vector notation for virial stress instead of individual position vectors.
In order to remove potential overhead, the functionality have been
involved in the analysis code in time_evolution library, with reference
variables.

Verification for dumbbell model perfectly matched with its theoretical
expectation. Verification with stochastic_HEUR_simulation have been
visited, which is not so much differ from the existing one. The
quantitative comparison with the existing one will be re-visit soon.


Former-commit-id: 55a423b3bf4de1db7e10a97e837b27043cafdeb2