[MAJOR UPDATE][IMPORTANT] CUTOFF for ASSOCIATION/DISSOCIATION From now one, the cutoff_connection work both of association and dissociation scheme. Whenever the bridge length larger than cutoff_connection, it will be check during topological update, then destroy. Note that this cutoff_connection only works with "Modified_Gaussian" and "FENE" options for both of equilibrium and non-equilibrium simulation.
The main idea of this cutoff_connection update is that we have limited
box size while the strong shear rate introduce very stretched chain
during non-equilibrium simulation. When the bridge length is larger than
half of box dimeter, the measured distance is change to some other image
set because of minimum image convention. Tracking all the chain on this
aspect is not easy for the current code architecture, while the
increasing box size increses computation time which is un-desirable
case. Hence, we decided to introduce artificial cutoff. Of course, the
FENE case is reasonable since the bridge length cannot be larger than
maximally extendable chain. Even though, the previous simulation code
does not accout for destroy (only association has the information),
which has problem when the shear rate is larger. With this update,
thouse problem reported on here will be solved.

Guideline: using cutoff_connection with a certain value no matter what
the cell_list option is turned on or not. This value must be lower than
Ld/2, but need a little further decreasing due to the a certain
tolerance. For instance, Wi_R=10^1 for simple shear with
box_dimension=10, if time step is crudely choose as dt=10^-2, the top
shift factor is about 1, which will increase length of bridge quite
larger. The factor should be smaller than 1. In the case with Wi_R=10^2
with box_dimension=10, the dt=10^-2 is too small. If a bridge perfectly
lies on shear gradient direction with length of 1, the length became
sqrt(2)~1.414. It means if the chain length in shear gradient direction
larger than 2, the changes is about 2.8 which is quite big. So keep
aware about shear rate region and also time step.

Former-commit-id: 355ca2068d6387dc27f4fb865eca86a7b0f89f49