[UPDATE] repulsion_type and repulsion_coupling options for inp file In this update, new architecture for repulsive interaction is applied. The option is called 'repulsion_type' and 'repulsion_coupling'.
Before going further, the basic potential type will be described on here. First of all, the only soft-repulsion type is following the form of
f(r) = C_rep * (1- r^2) when r<1.
The problem is that which type of C_rep.

In order to keep the compatibility of source code in the previous version, without the specific type of repulsion_type and repulsion_coupling, the C_rep is set with the given value of input file in the 'repulsion_coefficient'. It means we give C_rep value as pre-averaged repulsive coefficient value, then keeps all the coupled repulsion using the same C_rep.

In the revised version, we can select variety type for repulsion_type. The test function is called 'SOFT_REPULSION_P2' where P2 means C_rep is proportional to the p^2 where p is aggregation number of a micelle, which means C_rep depends on the number of chain ends of a specific micelle. Therefore, it is prefer to define C_i = C_0 p_i^2 where index i is for micelle index. In this way, we have to specify how to couple between different repulsive potential, since the applied repulsive interaction is pair-wise additivity. The answer is 'simple geometrical mean' way at this moment, but it would be better to open the other coupling methods for the future. Hence, we define following two input file arguments:
(i) repulsion_type (the only one specific option is SOFT_REPULSION_P2. Otherwise, it will use the default function);
(ii) repulsion_coupling (there are two options: PRE_AVERAGE and GEOMETRIC_MEAN).

Note that the repulsion_type=SOFT_REPULSION_P2 and repulsion_coupling=PRE_AVERAGE will be the same as the previous default test because it gives by the value C_rep as input file of repulsion_coefficient. If repulsion_coupling=GEOMETRICAL_MEAN< but repulsion_type is the default one, it also does not change anything. At this moment, the only repulsion_type=SOFT_REPULSION_P2 "and" repulsion_coupling=GEOMETRIC_MEAN will apply the thing. Even so, it is worth to decouple for future work.


Former-commit-id: cc9f7c241a62a1c36be92a141dbbad0116be8670